requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1
boto3==1.35.88
botocore==1.35.88
python-multipart==0.0.17
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1
Pillow==11.0.0
numpy==2.2.0
nova-act==0.2.0
